requirements.txt
requests
pyyaml
beautifulsoup4
psycopg2-binary
python-dotenv
asyncpg
aiohttp
